Refactor featured cards
