--- src/components/cards/featured-medium/types.ts
@@ -1,12 +1,0 @@
-import { PillCategory } from '../../../types/pill-category';
-import { Thumbnail } from '../../../interfaces/thumbnail';
-
-export interface FeaturedMediumCardProps {
-    contentDate: string;
-    className?: string;
-    description?: string;
-    title: string;
-    pillCategory: PillCategory;
-    tags?: string[];
-    thumbnail?: Thumbnail;
-}
