--- src/components/cards/index.ts
@@ -1,4 +1,0 @@
-import FeaturedMedium from './featured-medium/featured-medium';
-import FeaturedLarge from './featured-large/featured-large';
-
-export { FeaturedMedium, FeaturedLarge };
