--- src/components/cards/utils.ts
@@ -1,5 +1,0 @@
-import { ImageLoaderProps } from 'next/dist/client/image';
-
-export const thumbnailLoader = ({ src, width, quality }: ImageLoaderProps) => {
-    return `${src}?w=${width}&q=${quality || 75}`;
-};
